Remove warning for conversion from double to foat for akaze, and fix doc of BRISK

+++ modules/features2d/misc/java/test/FREAKDescriptorExtractorTest.java
@@ -46,7 +46,7 @@ public class FREAKDescriptorExtractorTest extends OpenCVTestCase {
 
         assertEquals(false, extractor.getOrientationNormalized());
         assertEquals(false, extractor.getScaleNormalized());
-        assertEquals((float) 23.0, extractor.getPatternScale());
+        assertEquals(23.0, extractor.getPatternScale());
         assertEquals(5, extractor.getNOctaves());
     }
 
@@ -58,7 +58,7 @@ public class FREAKDescriptorExtractorTest extends OpenCVTestCase {
 
         assertEquals(false, extractor.getOrientationNormalized());
         assertEquals(false, extractor.getScaleNormalized());
-        assertEquals((float) 23.0, extractor.getPatternScale());
+        assertEquals(23.0, extractor.getPatternScale());
         assertEquals(5, extractor.getNOctaves());
     }
